body_extraction/requirements.txt
dnspython
pandas
lxml
Cython
psycopg2-binary
numpy
dragnet
